Add files via upload
BrakeGuard is a comprehensive web based platform that helps drivers stay informed about the condition of
their brakes in real-time.BrakeGuard incorporates a user-friendly interface built with HTML, CSS, and JavaScript
(JS) to provide drivers an intuitive and visually appealing experience. The use of Node.js and WebSockets enables real-time data communication between the website and the brake monitoring system. Leveraging a data
visualization library, the website presents the information in a clear and interactive manner, allowing drivers to
easily interpret the data. The integration of MySQL ensures secure and efficient storage of the collected data,
enabling historical analysis and insights